refactor: record/[id] 페이지 SSR 전환 및 Next.js 15 에러 해결
- record/[page.tsx]를 서버 컴포넌트로 리팩터링하여 초기 데이터(LCP) 렌더링 속도 최적화
- 기존 클라이언트 측 로직은 [RecordDetailClient.tsx]로 관심사 분리
- Next.js 15 규격(Breaking Change)에 맞춰 `params`, `searchParams` Props 비동기(Promise) 처리 적용
