+++ src/components/record/RecordImageCarousel.tsx
@@ -141,6 +141,7 @@ export const RecordImageCarousel = ({
                 src={src}
                 alt={`Record image ${index + 1}`}
                 fill
+                quality={60}
                 className="object-cover pointer-events-none"
                 style={{ transform: `scale(${scale})` }}
                 priority={isFirstRecord && index === currentIndex}
